Issuse175 (#264)
* added new components for different function

* added new Scss for their following component

* Changed the exports for components

* import correct componets path

* imported prtotypes

* added new line

* imported react

* test

* added scss in index.js

* added new components

* chagned card name
