+++ app/frontend/src/components/Cards/InnerCopCard.tsx
@@ -1,0 +1,17 @@
+import { combineClasses } from "../Utility/utils";
+
+import React from "react";
+
+interface InnerCopCardProps extends React.PropsWithChildren {
+  addClass?: string;
+}
+
+function InnerCopCard(props: InnerCopCardProps) {
+  return (
+    <div className={combineClasses("card", "inner-cop-card", props.addClass)}>
+      <div className="inner-cop-card-content">{props.children}</div>
+    </div>
+  );
+}
+
+export { InnerCopCard };
