+++ app/frontend/src/components/Cards/_CopCard.scss
@@ -76,44 +76,3 @@
     left: 6.4vw;
   }
 }
-
-/**********************************
-*** Classes for inner COP cards ***
-**********************************/
-.inner-cop-card {
-  background-color: $color-white;
-  box-shadow: 0 8px 8px rgb(0 0 0 / 25%);
-  height: 100%;
-  max-width: 625px;
-  min-height: 448px;
-}
-
-.inner-cop-card-content {
-  box-sizing: border-box;
-  height: 100%;
-  padding: {
-    top: 48px;
-    right: 3.3vw;
-    left: 3.3vw;
-    bottom: 48px;
-  }
-}
-
-/**************************************
-*** Classes for inner COP nav cards ***
-**************************************/
-.inner-cop-nav-card {
-  background-color: $color-white;
-  border: none;
-  max-width: 188px;
-  min-height: 64px;
-
-  &.active {
-    background-color: $color-blue-dark;
-    color: $color-white;
-  }
-
-  &:hover {
-    cursor: pointer;
-  }
-}
