+++ app/frontend/src/components/Cards/_InnerCopCard.scss
@@ -1,0 +1,21 @@
+// Internal Imports
+@use "../Basics" as *;
+
+.inner-cop-card {
+  background-color: $color-white;
+  box-shadow: 0 8px 8px rgb(0 0 0 / 25%);
+  height: 100%;
+  max-width: 625px;
+  min-height: 448px;
+}
+
+.inner-cop-card-content {
+  box-sizing: border-box;
+  height: 100%;
+  padding: {
+    top: 48px;
+    right: 3.3vw;
+    left: 3.3vw;
+    bottom: 48px;
+  }
+}
