+++ app/frontend/src/index.scss
@@ -6,7 +6,9 @@
 
 // Import component styles
 @use "components/Buttons" as *;
-@use "components/Cards/Cards";
+@use "components/Cards/CopCard";
+@use "components/Cards/InnerCopCard";
+@use "components/Cards/InnerCopNavCard";
 @use "components/Cards/CircleCard";
 @use "components/Carousel/ScrollCarousel";
 @use "components/Dialog/Dialog";
